first draft done

+++ Backend/controllers/upload.js
@@ -6,40 +6,41 @@ const path = require('path');
 
 module.exports = async (req, res) => {
   try {
-    //Create unique file name and save in db
-    const uuid = uuidv4();
+    //Contoll file size
+    if (req.files.file.size / (1024 * 1024) < 5) {
+      //Create unique file name and save in db
+      const uuid = uuidv4();
 
-    const upload = new Upload({
-      fileName: uuid,
-    });
-
-    await upload.save();
-
-    //if not .zip => Convert files to zip and then save temporary to server directory
-    const output = fs.createWriteStream(__dirname + '/../output/' + uuid + '.zip');
-    const archive = archiver('zip');
-
-    if (path.extname(req.files.file.name) !== '.zip') {
-      archive.on('error', async (err) => {
-        await Upload.findOneAndDelete({ fileName: uuid });
-        throw err;
+      const upload = new Upload({
+        fileName: uuid,
       });
 
-      archive.pipe(output);
+      await upload.save();
 
-      const buffer = Buffer.from(req.files.file.data);
-      archive.append(buffer, { name: `${req.files.file.name}` });
+      //if not .zip => Convert files to zip and then save temporary to server directory
+      const output = fs.createWriteStream(__dirname + '/../output/' + uuid + '.zip');
+      const archive = archiver('zip');
 
-      await archive.finalize();
+      if (path.extname(req.files.file.name) !== '.zip') {
+        archive.on('error', async (err) => {
+          await Upload.findOneAndDelete({ fileName: uuid });
+          throw err;
+        });
+
+        archive.pipe(output);
+
+        const buffer = Buffer.from(req.files.file.data);
+        archive.append(buffer, { name: `${req.files.file.name}` });
+
+        await archive.finalize();
+      } else {
+        req.files.file.mv(path.resolve(__dirname + '/../output/' + uuid + '.zip'));
+      }
+      //End response
+      res.status(201).json({ data: uuid });
     } else {
-      req.files.file.mv(path.resolve(__dirname + '/../output/' + uuid + '.zip'));
+      res.status(406).json({ message: 'File size exceed 5mb!' });
     }
-
-    //Send files to cloud storage
-    console.log(archive.pointer());
-
-    //End response
-    res.status(200).json({ data: uuid });
   } catch (err) {
     console.log(err);
     res.status(500).end();
